+++ AudioMixer/UI/Root/Mixer.swift
@@ -172,70 +172,6 @@ struct Mixer: View {
     }
 }
 
-// MARK: - RecordSelectionView (unchanged)
-
-struct RecordSelectionView: View {
-    @ObservedObject private var viewModel = RecordAudioViewModel()
-    var onSelect: (URL) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("Select file record")
-                .font(.title2)
-                .fontWeight(.semibold)
-                .padding(.bottom, 8)
-            if viewModel.recordings.isEmpty {
-                VStack(spacing: 16) {
-                    Spacer()
-                    Image(systemName: "waveform")
-                        .resizable()
-                        .frame(width: 48, height: 48)
-                        .foregroundColor(.gray)
-                    Text("There is no recording")
-                        .foregroundColor(.gray)
-                        .font(.body)
-                    Spacer()
-                }
-            } else {
-                List {
-                    ForEach(viewModel.recordings) { recording in
-                        Button(action: {
-                            onSelect(recording.fileURL)
-                        }) {
-                            HStack {
-                                Image(systemName: "waveform")
-                                    .foregroundColor(.accentColor)
-                                VStack(alignment: .leading) {
-                                    Text(recording.fileURL.lastPathComponent)
-                                        .font(.headline)
-                                    Text(Self.dateFormatter.string(from: recording.createdAt))
-                                        .font(.caption)
-                                        .foregroundColor(.gray)
-                                }
-                                Spacer()
-                            }
-                            .padding(.vertical, 8)
-                        }
-                    }
-                }
-                .listStyle(.plain)
-            }
-        }
-        .padding()
-        .onAppear {
-            viewModel.fetchRecordings()
-        }
-    }
-    
-    private static let dateFormatter: DateFormatter = {
-        let df = DateFormatter()
-        df.dateStyle = .medium
-        df.timeStyle = .short
-        return df
-    }()
-}
-
-// MARK: - MusicFile Model & PlayMusicViewModel
 
 struct MusicFile: Identifiable {
     let id = UUID()
@@ -244,94 +180,8 @@ struct MusicFile: Identifiable {
     let duration: String
 }
 
-class PlayMusicViewModel: ObservableObject {
-    @Published var musicFiles: [MusicFile] = []
-
-    func fetchMusicFiles() {
-        musicFiles.removeAll()
-        let fileManager = FileManager.default
-        let documents = fileManager.urls(for: .documentDirectory, in: .userDomainMask).first!
-
-        // Find all mp3 files in Documents directory
-        let files = (try? fileManager.contentsOfDirectory(at: documents, includingPropertiesForKeys: nil)) ?? []
-        for file in files where file.pathExtension.lowercased() == "mp3" {
-            let asset = AVURLAsset(url: file)
-            let durationSeconds = CMTimeGetSeconds(asset.duration)
-            let minutes = Int(durationSeconds) / 60
-            let seconds = Int(durationSeconds) % 60
-            let durationString = String(format: "%d:%02d", minutes, seconds)
-            let displayName = file.deletingPathExtension().lastPathComponent
-            musicFiles.append(MusicFile(fileURL: file, displayName: displayName, duration: durationString))
-        }
-        // For demo: add a bundled "demo" if present
-        if let demoURL = Bundle.main.url(forResource: "demo", withExtension: "mp3") {
-            let asset = AVURLAsset(url: demoURL)
-            let durationSeconds = CMTimeGetSeconds(asset.duration)
-            let minutes = Int(durationSeconds) / 60
-            let seconds = Int(durationSeconds) % 60
-            let durationString = String(format: "%d:%02d", minutes, seconds)
-            musicFiles.append(MusicFile(fileURL: demoURL, displayName: "Demo", duration: durationString))
-        }
-        // Sort by displayName
-        musicFiles.sort { $0.displayName < $1.displayName }
-    }
-}
-
 // MARK: - PlayMusicSelectionView
 
-struct PlayMusicSelectionView: View {
-    @ObservedObject private var viewModel = PlayMusicViewModel()
-    var onSelect: (URL) -> Void
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("Chọn file nhạc đã lưu")
-                .font(.title2)
-                .fontWeight(.semibold)
-                .padding(.bottom, 8)
-            if viewModel.musicFiles.isEmpty {
-                VStack(spacing: 16) {
-                    Spacer()
-                    Image(systemName: "music.note.list")
-                        .resizable()
-                        .frame(width: 48, height: 48)
-                        .foregroundColor(.gray)
-                    Text("Không có file nhạc nào")
-                        .foregroundColor(.gray)
-                        .font(.body)
-                    Spacer()
-                }
-            } else {
-                List {
-                    ForEach(viewModel.musicFiles) { file in
-                        Button(action: {
-                            onSelect(file.fileURL)
-                        }) {
-                            HStack {
-                                Image(systemName: "music.note")
-                                    .foregroundColor(.accentColor)
-                                VStack(alignment: .leading) {
-                                    Text(file.displayName)
-                                        .font(.headline)
-                                    Text(file.duration)
-                                        .font(.caption)
-                                        .foregroundColor(.gray)
-                                }
-                                Spacer()
-                            }
-                            .padding(.vertical, 8)
-                        }
-                    }
-                }
-                .listStyle(.plain)
-            }
-        }
-        .padding()
-        .onAppear {
-            viewModel.fetchMusicFiles()
-        }
-    }
-}
 
 #Preview {
     Root()
